Lib: Refactor ws user search module creation
