+++ src/routes/(platform)/dashboard/task/[taskUuid]/update/+page.svelte
@@ -9,23 +9,14 @@
     import {
         assignLabelToTask,
         updateTask as performUpdateTask,
-        assignUserToTask,
     } from "$lib/repository/workspace";
-    import {
-        currentWorkspaceUsers,
-        createWorkspaceUserSearchResults,
-    } from "$lib/stores/dashboard";
+    import { createWorkspaceUserSearchModule } from "$lib/stores/dashboard";
     import { openContextMenu } from "$lib/stores/globalUi";
     import { createLabelSearchModule } from "$lib/stores/modules";
     import type {
         WorkspaceUserSearchModule,
         TaskModule,
     } from "$lib/types/stores";
-    import type {
-        TasksPerUser,
-        WorkspaceUserSelection,
-        WorkspaceUserSelectionInput,
-    } from "$lib/types/ui";
     import type { Task } from "$lib/types/workspace";
 
     export let data: { task: Task };
@@ -53,56 +44,8 @@
     }
 
     $: {
-        // Only do this once in the beginning
-        const { workspace_board_section: workspaceBoardSection } = task;
-        if (!workspaceBoardSection) {
-            throw new Error("Expected workspaceBoardSection");
-        }
-        const { workspace_board: workspaceBoard } = workspaceBoardSection;
-        if (!workspaceBoard) {
-            throw new Error("Expected workspaceBoard");
-        }
-        const { workspace } = workspaceBoard;
-        if (!workspace) {
-            throw new Error("Expected workspace");
-        }
-        const workspaceUserSearch = writable("");
-        const selected: WorkspaceUserSelection = task.assignee
-            ? {
-                  kind: "workspaceUsers",
-                  workspaceUserUuids: new Set([task.assignee.uuid]),
-              }
-            : {
-                  kind: "unassigned",
-              };
-        const workspaceUserSearchModule: WorkspaceUserSearchModule = {
-            select: async (selection: WorkspaceUserSelectionInput) => {
-                if (selection.kind === "unassigned") {
-                    await assignUserToTask(null, task.uuid);
-                } else if (selection.kind === "allWorkspaceUsers") {
-                    throw new Error("Unsupported");
-                } else {
-                    await assignUserToTask(
-                        selection.workspaceUser.user.email,
-                        task.uuid
-                    );
-                }
-            },
-            deselect: console.error,
-            selected: writable<WorkspaceUserSelection>(selected),
-            // XXX find a way to postpone this, albeit useful, showing
-            // the amount of tasks per users right from the beginning
-            // will be more work
-            tasksPerUser: readable<TasksPerUser>({
-                unassigned: 0,
-                assigned: new Map(),
-            }),
-            search: workspaceUserSearch,
-            searchResults: createWorkspaceUserSearchResults(
-                currentWorkspaceUsers,
-                workspaceUserSearch
-            ),
-        };
+        const workspaceUserSearchModule: WorkspaceUserSearchModule =
+            createWorkspaceUserSearchModule(task);
         const labelSearchModule = createLabelSearchModule(
             task,
             async (labelUuid: string, selected: boolean) => {
